weather/requirements.txt
Django==4.2
requests==2.31.0
psycopg2-binary
python-decouple==3.8